+++ CHANGELOG.md
@@ -10,10 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Pushed `pysui` 0.1.1 to [PyPi](https://pypi.org/project/pysui/)
+- Support for `sui_splitCoinEqual`
 
 ### Fixed
 
 ### Changed
+- **SuiClient** and **SuiAsynchClinet** [issue](https://github.com/FrankC01/pysui/issues/27)
 
 ### Removed